fix CategoryServiceImpl, CompilationServiceImpl, EventServiceImpl

+++ main-service/main-dto/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -1,11 +1,11 @@
 package ru.practicum.dto.event;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import ru.practicum.dto.category.CategoryDto;
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
+import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.user.UserShortDto;
 
 import java.time.LocalDateTime;
@@ -20,12 +20,12 @@ public class EventShortDto {
     String annotation;
     @NotNull
     CategoryDto category;
-    int confirmedRequests;
+    long confirmedRequests;
     @NotNull
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @NotNull
     LocalDateTime eventDate;
-    int id;
+    long id;
     @NotNull
     UserShortDto initiator;
     @NotNull
